requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
cryptography==41.0.7
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
stripe==7.8.0
requests==2.31.0
Pillow==10.1.0
bleach==6.1.0
payfast-sdk==0.0.1.2
